Week 7 Day 4: Major compilation progress (88.5% complete)
Reduced errors from 104 to 12 (88.5% reduction)

Major fixes:
- Fixed all GPU ID type casting issues (u64 to u32)
- Fixed vault_cascade SecureKey usage
- Added all missing syscall match arms (19 new syscalls)
- Fixed borrow checker issue in syscall_advanced_ops (dup2)

Remaining: 12 errors
- GenericArray conversions (3)
- Thread safety markers (4)
- Borrow checker (3)
- Type mismatches (1)
- Type annotations (1)

Next: Fix remaining borrow checker and trait bound issues

+++ src/verified/direct_metal_metal.rs
@@ -337,7 +337,7 @@ impl GpuBackend for MetalBackend {
             library: None,
         };
         
-        let device_id = self.next_device_id;
+        let device_id = self.next_device_id as u32;
         self.next_device_id += 1;
         
         self.devices.insert(device_id, device);
@@ -436,7 +436,7 @@ impl GpuBackend for MetalBackend {
             current_encoder: None,
         };
         
-        let cmd_id = self.next_cmd_id;
+        let cmd_id = self.next_cmd_id as u32;
         self.next_cmd_id += 1;
         
         self.command_buffers.insert(cmd_id, cmd_buffer);
@@ -532,7 +532,7 @@ impl GpuBackend for MetalBackend {
             command_buffer: None,
         };
         
-        let fence_id = self.next_fence_id;
+        let fence_id = self.next_fence_id as u32;
         self.next_fence_id += 1;
         
         self.fences.insert(fence_id, fence);
@@ -587,7 +587,7 @@ impl GpuBackend for MetalBackend {
             compute_pipeline_state: compute_state,
         };
         
-        let pipeline_id = self.next_pipeline_id;
+        let pipeline_id = self.next_pipeline_id as u32;
         self.next_pipeline_id += 1;
         
         self.pipelines.insert(pipeline_id, pipeline);
